twitter_trend_view/send_mail/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud
google-cloud-storage
pytz

